+++ server/routes/api/v1/db.py
@@ -1,0 +1,9 @@
+"""Implements database routes"""
+
+import logging
+
+from fastapi import APIRouter
+
+logger = logging.getLogger(__name__)
+
+router = APIRouter()