+++ server/routes/api/v1/room.py
@@ -1,0 +1,9 @@
+"""Implements room routes"""
+
+import logging
+
+from fastapi import APIRouter
+
+logger = logging.getLogger(__name__)
+
+router = APIRouter()