+++ tests/test_root.py
@@ -1,0 +1,18 @@
+"""Unit tests for the root routes"""
+
+from fastapi import status
+
+
+def test_get_root(client):
+    """Test the hello world response"""
+
+    response = client.get('/')
+    assert response.status_code == status.HTTP_200_OK, response.text
+    assert response.json() == {"hello": "Hello, world!"}, response.json()
+
+
+def test_get_robots_txt(client):
+    """Test getting the robots.txt data"""
+
+    response = client.get('/robots.txt')
+    assert response.status_code == status.HTTP_200_OK, response.text